Add JSON extra field type
The extra field stores an arbitrary JSON subobject as a field on the
root extra_fields object. This is intended for automated systems to fill
out and provide data for templating. Since it's likely to contain a
large value, Tte UI hides the value behind a button, like rich text
fields. It also uses a plain non-rich-text textarea for editing.